feat(scripts): use snapshots for droplet
This will cost more, but will make devserver more useful.

When destroying the droplet, the client will take a snapshot of the
droplet, delete any previous snapshots, and then destroy the droplet.

When creating the droplet, the client will check if a snapshot exists.
If a snapshot exists, a new droplet will be created from it. Otherwise,
a new droplet will be created from the docker image.

+++ scripts/destroy_droplet.py
@@ -2,10 +2,18 @@ import sys
 import digital_ocean_client
 
 def destroy_droplet(client):
-    droplet_id = client.get_droplet(client.droplet_name)
-
+    droplet_info = client.get_droplet(client.droplet_name)
+    if droplet_info:
+        droplet_id = droplet_info.get('id')
+    else:
+        sys.exit(1)
+    if not droplet_id:
+        sys.exit(1)
+    print(f'Droplet ID:{droplet_id}')
     client.power_off(droplet_id)
     client.wait_for_droplet(droplet_id, False)
+    snapshot_name = client.snapshot(droplet_id)
+    client.clean_snapshots(snapshot_name)
     client.delete(droplet_id)
     print('Droplet destroyed')
